# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sutradhar API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/items": {
      "post": {
        "tags": [
          "provenance"
        ],
        "summary": "Register Item",
        "description": "Register an item. Item, event, and outbox row commit together or not at all.",
        "operationId": "register_item_api_v1_items_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterItemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "provenance"
        ],
        "summary": "List Items",
        "description": "Keyset page over items, newest first.",
        "operationId": "list_items_api_v1_items_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Slug"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ItemStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "registered_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Registered By"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/items/{item_id}/attestations": {
      "post": {
        "tags": [
          "attestation"
        ],
        "summary": "Record an attestation about an item",
        "description": "Vouch for an item. One attestation per actor per item, enforced by the database.\n\nA second attempt is a 409 from ``uq_attestations_item_attestor``, not from\nan application check: a check-then-write has a gap, and two simultaneous\nrequests would both pass it.",
        "operationId": "create_attestation_api_v1_items__item_id__attestations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttestationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttestationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "attestation"
        ],
        "summary": "List the attestations recorded against an item",
        "description": "Newest first. Carries roles and pseudonymous references, never identities.",
        "operationId": "list_attestations_api_v1_items__item_id__attestations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttestationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{item_id}/trust": {
      "get": {
        "tags": [
          "attestation"
        ],
        "summary": "Who vouched for this item, and how independent were they",
        "description": "Derived on every read from the attestation and dispute sets.\n\nNothing is cached and nothing is stored, so a fraud flag applied a second\nago is already reflected here.",
        "operationId": "get_trust_api_v1_items__item_id__trust_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/actors/{user_id}/fraud-flag": {
      "post": {
        "tags": [
          "attestation",
          "admin"
        ],
        "summary": "Flag an actor and dispute everything they registered",
        "description": "One transaction: the flag, the audit event, and every affected item.\n\nItems the actor *registered* become disputed. Items they merely attested to\nare not touched -- their trust level drops on the next read, because the\nlevel is derived and a flagged attestor stops counting.",
        "operationId": "fraud_flag_api_v1_admin_actors__user_id__fraud_flag_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagActorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagActorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/actors/{user_id}/fraud-clear": {
      "post": {
        "tags": [
          "attestation",
          "admin"
        ],
        "summary": "Lift a fraud flag and the disputes it caused",
        "description": "Reverses exactly what the flag did, and nothing else.\n\nAn item independently disputed by an inspector stays disputed: that finding\nis not this flag's to lift.",
        "operationId": "fraud_clear_api_v1_admin_actors__user_id__fraud_clear_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagActorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagActorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttestationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttestationResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AttestationListResponse",
        "description": "A page of attestations, newest first."
      },
      "AttestationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "attestor_ref": {
            "type": "string",
            "title": "Attestor Ref"
          },
          "attestor_role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "attestor_fraud_flagged": {
            "type": "boolean",
            "title": "Attestor Fraud Flagged"
          },
          "statement": {
            "additionalProperties": true,
            "type": "object",
            "title": "Statement"
          },
          "statement_hash": {
            "type": "string",
            "title": "Statement Hash"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "item_id",
          "attestor_ref",
          "attestor_role",
          "attestor_fraud_flagged",
          "statement",
          "statement_hash",
          "created_at"
        ],
        "title": "AttestationResponse",
        "description": "One recorded attestation, with the attestor pseudonymised."
      },
...
      "CreateAttestationRequest": {
        "properties": {
          "statement": {
            "additionalProperties": true,
            "type": "object",
            "title": "Statement",
            "description": "Free-form claim: inspection notes, ledger references, dates."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "statement"
        ],
        "title": "CreateAttestationRequest",
        "description": "A claim one party makes about one item."
      },
...
      "DisputeStatus": {
        "type": "string",
        "enum": [
          "NONE",
          "DISPUTED"
        ],
        "title": "DisputeStatus",
        "description": "Whether an item's provenance is contested."
      },
      "FlagActorRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 8,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "FlagActorRequest",
        "description": "Admin action. The reason is required and is shown on every item it touches."
      },
      "FlagActorResponse": {
        "properties": {
          "actor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Actor Id"
          },
          "fraud_flagged": {
            "type": "boolean",
            "title": "Fraud Flagged"
          },
          "items_affected": {
            "type": "integer",
            "title": "Items Affected"
          },
          "attestations_affected": {
            "type": "integer",
            "title": "Attestations Affected"
          },
          "already_in_state": {
            "type": "boolean",
            "title": "Already In State",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "actor_id",
          "fraud_flagged",
          "items_affected",
          "attestations_affected"
        ],
        "title": "FlagActorResponse",
        "description": "What a flag or a clear actually changed, reported back plainly."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ItemSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "additionalProperties": true,
            "type": "object",
            "title": "Pagination"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "ItemListResponse",
        "description": "Collection envelope with a keyset cursor, per app.core.pagination."
      },
...
      "ItemStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "CONFIRMED",
          "FAILED"
        ],
        "title": "ItemStatus",
        "description": "Anchoring state of an item's hash. PENDING until the chain confirms."
      },
      "ItemSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "category_schema_version": {
            "type": "integer",
            "title": "Category Schema Version"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "registered_by": {
            "type": "string",
            "format": "uuid",
            "title": "Registered By"
          },
          "quantity": {
            "type": "string",
            "title": "Quantity"
          },
          "quantity_unit": {
            "type": "string",
            "title": "Quantity Unit"
          },
          "item_hash": {
            "type": "string",
            "title": "Item Hash"
          },
          "tag_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Code"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "dispute_status": {
            "$ref": "#/components/schemas/DisputeStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "category_schema_version",
          "parent_id",
          "registered_by",
          "quantity",
          "quantity_unit",
          "item_hash",
          "tag_code",
          "status",
          "dispute_status",
          "created_at"
        ],
        "title": "ItemSummary",
        "description": "An item in a listing."
      },
...
      "RegisterItemRequest": {
        "properties": {
          "category_slug": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Category Slug"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Quantity"
          },
          "quantity_unit": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Quantity Unit"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "type": "object",
        "required": [
          "category_slug",
          "attributes",
          "quantity",
          "quantity_unit"
        ],
        "title": "RegisterItemRequest",
        "description": "Register a new item, or a root of a new provenance tree."
      },
...
      "TrustLevel": {
        "type": "string",
        "enum": [
          "SELF_DECLARED",
          "CO_OP_ATTESTED",
          "INSPECTED",
          "DISPUTED"
        ],
        "title": "TrustLevel",
        "description": "How much independent corroboration a record has attracted.\n\nDeliberately a description of the evidence, not a judgement about the\nobject. None of these values says anything is real, and none of them ever\nshould -- see the module docstring."
      },
      "TrustResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "level": {
            "$ref": "#/components/schemas/TrustLevel"
          },
          "contributing_roles": {
            "items": {
              "$ref": "#/components/schemas/UserRole"
            },
            "type": "array",
            "title": "Contributing Roles"
          },
          "attestation_count": {
            "type": "integer",
            "title": "Attestation Count"
          },
          "distinct_attestor_count": {
            "type": "integer",
            "title": "Distinct Attestor Count"
          },
          "dispute_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispute Reason"
          },
          "flagged_attestor_count": {
            "type": "integer",
            "title": "Flagged Attestor Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "level",
          "contributing_roles",
          "attestation_count",
          "distinct_attestor_count"
        ],
        "title": "TrustResponse",
        "description": "What corroboration a record has attracted. Not a verdict about the object.\n\n``level`` describes the evidence, and the fields beneath it are that\nevidence, so a reader who wants to weigh it themselves can. A payload that\ncarried only the level would be asking to be trusted; this one shows its\nworking."
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "CONSUMER",
          "WEAVER",
          "COOP_OFFICER",
          "INSPECTOR",
          "ADMIN"
        ],
        "title": "UserRole",
        "description": "Who someone is in the supply chain. One role per user."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}